requirements.txt
opencv-python
visdom
lpips
torchmetrics
dominate
numpy
scipy
pillow
unidecode
requests
tqdm
dlib
